names of attributes and types changed

+++ bil_abonnement/src/main/java/ita3/car_subscription/entity/Damage.java
@@ -10,7 +10,7 @@ public class Damage {
     private String error;
     private String errorType;
     private int numbersOfErrors;
-    private int pricePerError;
+    private double pricePerError;
 
     @ManyToOne
     @JoinColumn(name = "subscription_id", referencedColumnName = "id") //fremmednøgle
@@ -18,6 +18,7 @@ public class Damage {
     public Damage() {
 
     }
+
     public Damage(String error, String errorType, int numbersOfErrors, int pricePerError) {
         this.error = error;
         this.errorType = errorType;
@@ -58,11 +59,11 @@ public class Damage {
         this.numbersOfErrors = numbersOfErrors;
     }
 
-    public int getPricePerError() {
+    public double getPricePerError() {
         return pricePerError;
     }
 
-    public void setPricePerError(int pricePerError) {
+    public void setPricePerError(double pricePerError) {
         this.pricePerError = pricePerError;
     }
 }
